+++ src/main/java/calculator/Input.java
@@ -1,4 +1,4 @@
-package level0;
+package calculator;
 
 import java.util.Arrays;
 import java.util.HashSet;
