+++ src/main/java/calculator/Main.java
@@ -1,4 +1,4 @@
-package level0;
+package calculator;
 
 import java.io.BufferedReader;
 import java.io.IOException;
