+++ src/main/java/calculator/Number.java
@@ -1,4 +1,4 @@
-package level0;
+package calculator;
 
 import java.util.Objects;
