+++ src/main/java/lottery/exception/DuplicateLotteryNumberException.java
@@ -1,4 +1,4 @@
-package level1.exception;
+package lottery.exception;
 
 public class DuplicateLotteryNumberException extends RuntimeException {
