+++ src/main/java/lottery/exception/InvalidLotteryNumberLengthException.java
@@ -1,4 +1,4 @@
-package level1.exception;
+package lottery.exception;
 
 public class InvalidLotteryNumberLengthException extends RuntimeException {
