+++ src/test/java/lottery/LotteryGeneratorTest.java
@@ -1,14 +1,14 @@
-package level1;
+package lottery;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Collections;
 import java.util.List;
-import level1.domain.AnswerLottery;
-import level1.domain.Lottery;
-import level1.exception.InvalidLotteryNumberLengthException;
-import level1.exception.InvalidLotteryNumberRangeException;
+import lottery.domain.AnswerLottery;
+import lottery.domain.Lottery;
+import lottery.exception.InvalidLotteryNumberLengthException;
+import lottery.exception.InvalidLotteryNumberRangeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
