+++ src/test/java/lottery/domain/LotteryTest.java
@@ -1,11 +1,11 @@
-package level1.domain;
+package lottery.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import java.util.stream.IntStream;
-import level1.exception.DuplicateLotteryNumberException;
+import lottery.exception.DuplicateLotteryNumberException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
